mentor_ai/requirements.txt
Django~=5.0.0               # Web framework
djangorestframework~=3.15.0 # REST API framework
djangorestframework-simplejwt~=5.3.1  # JWT authentication
psycopg2-binary~=2.9.9      # PostgreSQL database adapter
requests~=2.31.0            # News API requests
django-redis~=5.4.0         # Caching with Redis
openai~=1.12.0              # OpenAI API
httpx==0.27.2               # For async HTTP requests   
celery~=5.3.0               # Task queue
redis~=4.5.0                # Redis client
django-celery-beat~=2.6.0   # Periodic tasks with Celery
pgvector ~=0.4.0            # Vector storage in PostgreSQL
youtube-transcript-api~=1.2.0  # YouTube transcript extraction
